dashboard/requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==41.0.7
cycler==0.12.1
dill==0.3.8
distlib==0.3.7
Faker==24.4.0
filelock==3.12.3
fonttools==4.50.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaitaistruct==0.10
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
networkx==3.2.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pillow==10.3.0
pipenv==2023.8.28
platformdirs==3.10.0
playwright==1.40.0
plotly==5.20.0
podman-compose==1.0.6
protobuf==4.25.3
psycopg2==2.9.9
pyarrow==15.0.2
pyasn1==0.5.1
pycparser==2.21
pydeck==0.8.1b0
pydivert==2.1.0
pyee==11.0.1
Pygments==2.17.2
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
selenium==4.16.0
selenium-wire==5.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
streamlit==1.32.2
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
vboxapi==1.0
virtualenv==20.24.3
watchdog==4.0.0
webdriver-manager==4.0.1
wordcloud==1.9.3
wsproto==1.2.0
zstandard==0.22.0
